refactor: enhance document & knowledgebase modules with improved error handling and pagination
- Refactor document logic with better error handling and status checks
- Add document move, re-parse, preview and upload improvements
- Add document mapper, helper and pagination support
- Enhance knowledgebase logic with improved validation
- Update auth avatar upload with simplified flow
- Add document and knowledgebase repository improvements
- Add upload file utility and page repository
- Update router tests and request/response structures

+++ internal/transport/http/response/knowledgebase.go
@@ -20,8 +20,8 @@ type KnowledgeBaseResponse struct {
 	Color       string    `json:"color"`
 	IsDefault   bool      `json:"is_default"`
 	ChatEnabled bool      `json:"chat_enabled"`
-	DocCount    int       `json:"doc_count"`
-	ImageCount  int       `json:"image_count"`
+	DocCount    int64     `json:"doc_count"`
+	ImageCount  int64     `json:"image_count"`
 	CreatedAt   time.Time `json:"created_at"`
 	UpdatedAt   time.Time `json:"updated_at"`
 }
